Status button now functions
Status button functions using prototype function
according to steps in TOP.

Fixed an issue where the borders of the tds
under the options header had borders which were
too thick. Used a container div around the button
to fix this issue. (Display: flex in a td without
container seems to be what caused this.)

Fixed an issue where the tds under the status
header were resizing themselves when the content
changed. Added max-width style to resolve this.

Reflection:
Learnt a lot from this project. Spent a lot of
time polishing and learning new things. (Modals,
js prototype and constructors, usage of tables.)